Implement TrieMap type, with insertion, removal, finding, and basic iteration
Add a TrieMap implementation, that uses `LowMemoryThinVec` to store
node labels and children. This version includes implementations
for insertion, removal, finding, and ordered iteration over entries.
Additionally, `TrieMap::mem_usage` allows one to compute the current
memory usage of the trie.

Furthermore, an implementation of std::fmt::Debug is done,
as a means of inspecting the structure of the trie. Combined
with snapshot testing based on the `insta` crate, this allows
developers to verify the correctness of the data structure.

As a sanity check, a `proptest`-based tests is included,
which asserts the behavior of `TrieMap<T>` is identical
to that of `BTreeMap<Vec<c_char>, T>` given a randomized
input of operations.

The type of the `TrieMap` keys is `&[c_char]` to accomodate for
the fact that in C, `char` can either be unsigned or signed, depending
on the target platform, which affects ordering.

Co-authored-by: Luca Palmieri <20745048+LukeMathWalker@users.noreply.github.com>