Allow insertion in TrieMap given callback, use Redis allocator
This commit introduces TrieMap::insert_with, which allows
the caller to pass a callback function that gets passed the
current value corresponding to the key, if any, and return
the new value. This is integrated in `TrieMap::Add` to allow
the callback to decide which value should actually be inserted.

Furthermore, `redis-module` is added as a dependency,
and with it the Redis allocator is registered as the global
allocator in case the `redis_allocator` feature has been enabled,
which it is by default. `TrieMap::Add` now calls `rm_free` to
free any replaced data in case no callback is passed.