Implement `trie_rs::ffi::TrieMapResultBuf` based on `LowMemoryThinVec`
Replaces the empty `TrieMapResultBuf` enum with a newtype
that wraps a `LowMemoryThinVec<*mut c_void>`. Adds
a function TrieMapResultBuf_Len, allowing for obtaining
the length of the buffer, which until now could only
be obtained from the return value of `TrieMap_FindPrefixes`.

Furthermore, this commit implements `TrieMapResultBuf_Free`
and `TrieMapResultBuf_Data`