Enable statically linking to libclang when building bindgen
This commit updates `redis-module` to a fork
based on its master version that allows for statically
linking with libclang, a requirement on Alpine.
It also installs the packages needed to do so.

Furthermore, this commit conditionally enables statically
linking with libclang based on the target architecture
matching that of Alpine or not.

A flag is added to the make config, RUST_DYN_CRT,
which forces dynamically linking the C runtime,
and the flag is set for the Alpine platform on CI