[MOD-8972] - Implement basic TrieMap functionality in Rust (#5797)
* Implement LowMemoryThinVec::{mem_usage, prepend_with_slice, from_slice}

Adds methods `mem_usage`, `prepend_with_slice`, and `from_slice`
to `LowMemoryThinVec`.
`mem_usage` returns the size of the size of the heap allocation backing
the `LowMemoryThinVec`, `prepend_with_slice` prepends the held
items with a given slice, and `from_slice` creates a `LowMemoryThinVec`
from a slice, cloning the elements.

* Implement TrieMap type, with insertion, removal, finding, and basic iteration

Add a TrieMap implementation, that uses `LowMemoryThinVec` to store
node labels and children. This version includes implementations
for insertion, removal, finding, and ordered iteration over entries.
Additionally, `TrieMap::mem_usage` allows one to compute the current
memory usage of the trie.

Furthermore, an implementation of std::fmt::Debug is done,
as a means of inspecting the structure of the trie. Combined
with snapshot testing based on the `insta` crate, this allows
developers to verify the correctness of the data structure.

As a sanity check, a `proptest`-based tests is included,
which asserts the behavior of `TrieMap<T>` is identical
to that of `BTreeMap<Vec<c_char>, T>` given a randomized
input of operations.

The type of the `TrieMap` keys is `&[c_char]` to accomodate for
the fact that in C, `char` can either be unsigned or signed, depending
on the target platform, which affects ordering.

Co-authored-by: Luca Palmieri <20745048+LukeMathWalker@users.noreply.github.com>

* Forward extern fns in trie_rs::ffi to `TrieMap`.

Replaces the placeholder `TrieMap` in `trie_rs::ffi`
with a newtype, wrapping a `trie_rs::trie::TrieMap<*mut c_void>`.

Forward calls to `NewTrieMap`, `TrieMap_Add`, `TrieMap_Find`,
`TrieMap_Delete`, `TrieMap_Free`, and `TrieMap_MemUsage` to their
counterpart methods of `TrieMap<*mut c_void>`.

* Allow insertion in TrieMap given callback, use Redis allocator

This commit introduces TrieMap::insert_with, which allows
the caller to pass a callback function that gets passed the
current value corresponding to the key, if any, and return
the new value. This is integrated in `TrieMap::Add` to allow
the callback to decide which value should actually be inserted.

Furthermore, `redis-module` is added as a dependency,
and with it the Redis allocator is registered as the global
allocator in case the `redis_allocator` feature has been enabled,
which it is by default. `TrieMap::Add` now calls `rm_free` to
free any replaced data in case no callback is passed.

* Doc safety of memcpy call in `LowMemoryThinVec::prefix_with_slice`, resoning behind disabling snapshot tests when running under Miri

* Move dependency declarations to worspace manifest

* Clarify early return in Node::remove_child, don't merge child if it's a leaf

* Construct and insert all children directly in Node::insert_or_replace_with

* Move thinvec dependency definition to workspace manifest

* Disable proptest under miri

* Remove fmt::Debug bounds where not necessary

* Clarify Node::insert_or_replace_with, rename Node::merge_child to merge_child_if_possible

* Apply suggestions from code review

Co-authored-by: alonre24 <alon.reshef@redis.com>

* Assign empty slice if len == 0 in TrieMap_Delete

* rm_free value if no freeCB given in TrieMap_Delete

* Clarify Node data initialization at recursion

* Enable statically linking to libclang when building bindgen

This commit updates `redis-module` to a fork
based on its master version that allows for statically
linking with libclang, a requirement on Alpine.
It also installs the packages needed to do so.

Furthermore, this commit conditionally enables statically
linking with libclang based on the target architecture
matching that of Alpine or not.

A flag is added to the make config, RUST_DYN_CRT,
which forces dynamically linking the C runtime,
and the flag is set for the Alpine platform on CI

* Point redis-module to patches branch to eliminate refs to memfd

---------

Co-authored-by: Luca Palmieri <20745048+LukeMathWalker@users.noreply.github.com>
Co-authored-by: alonre24 <alon.reshef@redis.com>