Implement `TrieMap_Iterate, and TrieMapIterator_Next* in `trie_rs::ffi`
`TrieMap_Iterate` is implemented by applying different
iterator adaptors to `trie_rs::iter::LendingIter`, based
on the passed iteration mode. `TrieMapIterator_Next` simply
obtains the next value from the iterator, and
`TrieMapIterator_NextContains` and `TrieMapIterator_NextWildcard`
deletage to `TrieMapIterator_Next`.

This commit furthermore adds a number of tests
to validate the correct operation of the triemap
iterator functionality exposed in the `trie_rs::ffi` module.

This commit also removes the global allocator definition
set in `trie_rs`, as it hinders tests and is to be moved to a façade crate.