Introduce a facade crate, `redisearch_rs`. (#6011)
`redisearch_rs` is the unified entrypoint for C users of the Rust
code. It contains all the C-to-Rust FFI modules and it's the only
unit linked against the final redisearch.so library.

As part of the refactor, we also introduced `redis_mock`, a crate
that provides an alternative implementation of Redis' allocation
functions. It has been extracted from `trie_bencher` for future
use in other parts of the Rust workspace.