Start the redis-server via `runtests.sh`
By using `runtests.sh` we get coverage and a whole lot of other stuff
out of the box. The only issue that `runtests.sh` does not stop the
RLTest server it spawned. This requires a manual `SIGINT` to be send to
the `redis-server` (which RLTest spawns) to stop the RLTest server correctly.