Switch to `redis` crate
This allows more control over the parsing of the responses. This is
needed because `FT.SEARCH` does not have a set order for returning the
fields of each document. And since we used to only compare the string
responses, the comparison might fail just because the field orders are
different even though the field values are the same. Instead we are now
parsing these to a hash map and therefore the equality check works (when
the values are indeed the same) even if the Redis order of the fields
might not be the same.

This also has a nice benefit of speeding up the queries from running
1,500 queries in 20 seconds to running them in under a second.