[MOD-9798] Refactor `inverted_index` (#6220)
* Move write forward index entry to `forward_index.h`

This reduces the dependencies in `inverted_index.h` which will make it
easier to port to Rust.

The include rule in `rules.c` is needed to find `UNDERSCORE_PAYLOAD`
correctly. Else the build will fail.

* Remove unused definitions

All these definitions are never implemented or referenced. Removing them
means we won't have to be concerned about them when porting over the
`inverted_index.h` file.

* Make functions private

This makes functions which are only used by `inverted_index` private.
This will make the porting easier as these functions won't need to be in
the new C interface exposed by the Rust code.

* Add `IR_LastDocId` back for tests

The tests fail to compile without it.

* Update for CPP tests

The `InvertedIndex_WriteForwardIndexEntry` function is now in
`forward_index.h` instead.

* Simplify expression

Co-authored-by: GuyAv46 <47632673+GuyAv46@users.noreply.github.com>

---------

Co-authored-by: GuyAv46 <47632673+GuyAv46@users.noreply.github.com>