Remove unused dependencies from `src/value` header
Having the `<math.h>` in the `src/value.h` would cause the bindgen to
try and export some `boost` stuff. This would cause `FP_*` constants
like `FP_NORMAL` and `FB_SUBNORMAL` to be defined multiple times. Since
the header file does not use the math module it is same to move it down
to the definition file instead.

The other modules (`string`, `sys/param` and `stdarg`) are not used by
the `src/value` at all so are safely removed.