[MOD-9806] Index encoder & decoder interfaces (#6227)
* Make `inverted_index` crate library

* Make an `Encoder` interface for inverted index

This is meant to fullfill the same role as this C callback:

```c
/* An index encoder is a callback that writes records to the index. It accepts a pre-calculated
 * delta for encoding */
typedef size_t (*IndexEncoder)(BufferWriter *bw, t_docId delta, RSIndexResult *record);
```

* Switch to FFI types

* Make interface for Decoder

* Add `Seek` to interfaces

* Remove unused dependencies from `src/value` header

Having the `<math.h>` in the `src/value.h` would cause the bindgen to
try and export some `boost` stuff. This would cause `FP_*` constants
like `FP_NORMAL` and `FB_SUBNORMAL` to be defined multiple times. Since
the header file does not use the math module it is same to move it down
to the definition file instead.

The other modules (`string`, `sys/param` and `stdarg`) are not used by
the `src/value` at all so are safely removed.

* Move filter type to own module

This reduces the amount of things the bindgen generates and avoids
pulling in something (unknown) which causes a compile error.

* Add numeric filter to decoder context

* Remove decoder context

These context values can be stored on the structs implementing the
decoder trait instead.

* Make default implementation for seek

* Cleanup

Remove unused imports.

* Undo C changes

* Add `ffi` crate back

* Use explicit generics

* Improve documentation

* Add missing license header

* Cargo fmt

* Remove need for `Copy` trait

* Improve comment

Co-authored-by: Luca Palmieri <20745048+LukeMathWalker@users.noreply.github.com>

* Make Delta type

* Document delta

* Use a new type for `Delta` instead

* Improve comment

* Drop `EndOfStream` variant

It is replaced with an `Option` instead.

---------

Co-authored-by: Luca Palmieri <20745048+LukeMathWalker@users.noreply.github.com>