[MOD-9999] Migrate `RSAggregateResult` to Rust (#6255)
* Migrate `RSAggregateResult` to Rust

This is needed to move `RSIndexResult` to Rust which will make it easier
to port the inverted index to Rust.

* Better rename annotation

* Dummy `RSIndexResult` forward definition

This just makes the C compiler happy when "reading" values of the vector.

* Make fields public

* Better forward declaration

* Migrate `RSResultType` to Rust

This will makes the usages of `RSAggregateResult` more correct inside
the Rust code.

* Prefix enum name correctly

* Cast `int` to `RSResultType`

* Use a bit flag for the mask instead

* Use latest `enumflags2` until new release

The latest code has a fix we need for the Rust code to compile. But it
is not released yet so just using it directly for now.

* Rebase fixes

* Update `enumflags2` to latest release

* Add `enumflags2` to dev dependencies

This is needed to make the doc tests pass since the FFI is going to pull
in a comment which shows how to use `BitFlags`.