[MOD-9998] Migrate `RSIndexResult` to Rust (#6301)
* Migrate `RSIndexResult` to Rust

This is the last step to be able to more easily migrate the inverted
index to Rust too.

* Fix tests

* Reduce exports

* Remove `Copy` on types containing raw pointers

The problem with copy on raw pointers is that we might end up with
multiple mutable references to the raw types or just mixed mutable and
immutable references. This is obviously a memory / data race concert
which we don't want.

* Remove `Clone` since it is no longer needed

* Simplify derives on numeric record

Remove things which are no longer needed.