Use `with_capacity` instead
We should not want to grow the buffer for the inverted index tests and
benchmarks since that belongs in the buffer code. Instead we should
always make a buffer that is big enough for whatever is encoded into it.
This removes the `new` and add a `with_capacity` to make sure we always
use the test buffer in this way.