[MOD-10177] Rust benchmark for `numeric` encoder and decoder (#6336)
* Get a basic Rust benchmark working for the C code

This creates a Rust benchmark runner for the C numeric encoder and
decoder only. This will serve as a base for more encoder / decoder benchmarks.

* Don't make FFI for C buffer

The benchmarks are using the Rust wrapper instead.

* Make it easier to construct a buffer

* Prep for testing all the basic numeric inputs

* Test all the numeric encoding types

* Complete `Debug` and `PartialEq`

* Wrap up inverted index module

* Clean up the build file

* Finish up the C file changes

* Fix FFI include paths

This means `numeric_index.h` no longer needs a path to include `inverted_index.h`.

* Benchmark the encoding for all the types

* Use a bigger delta

This is make the benchmark more accurate.

* Benchmark numeric decoding

* Add missing license headers

* Add missing safety docs

* Fix includes for tests

* Fix spelling

* Add a README

* Rename `reset` to `clear`

* Fail if static library cannot be linked

* Use a buffer wrapper

This allows us to not worry about growth allocations and correctly
clears the memory after the buffer was used.

* Use `iter_batched`

* Remove the main

We won't need this as the output of the C and Rust should be exactly the
same. So there is no memory usage to compare.

* Fix CPP include

Missed fix for new file coming from master.

* Fix comment

* Better neg infinity constant

* Better name for buffer wrapper

* Use small input for batch size

* Ignore bencher during coverage

* Use `with_capacity` instead

We should not want to grow the buffer for the inverted index tests and
benchmarks since that belongs in the buffer code. Instead we should
always make a buffer that is big enough for whatever is encoded into it.
This removes the `new` and add a `with_capacity` to make sure we always
use the test buffer in this way.