[MOD-9801] Migrate numeric encoding/decoding to Rust (#6351)
* Simple encoder for tiny numeric values

* Also encode delta

* Make function to pack Delta into `u8` vec

* Test two byte delta

* Improve comments

* Implement numeric decoder

* Encode positive integers

* Decode positive ints

* Encode and decode negative ints

* Decode and encode floats

* Finish todos

* Make float value enum cleaner

* Simplify decoder matches

* Move to own module

* Make `ToFromBytes` trait

* Add Rust benchmarks

* Optimization pass

The Rust decoding is now on par with the C decoding. I still need to
understand why encoding is slower in most cases.

* Use vectored writes

The Rust is now faster at encoding than C in all cases.

* Use `copysign` to flip signs

This appears to give better performance than multiplying with -1.

* Improve numeric documentation

* Improve test documentation

* Make test buffer with big enough pre-allocted capacity

The idea is that none of the tests should need to call `BufferGrow`
since that is not what we are trying to profile anyway.

* More test groups

This makes a major group for each decoding / encoding type.

* Use match on all the float flags

This is about 14% faster than shifting bits.

* Do less bit shifts

This does a 10% improvement.

* Inline `pack` and `unpack`

This gives a 25% improvement.

* Better bench groups

This groups the benchmarks by the final bytes sizes decoded output.

* Add all delta ranges to the benches

* Group the benchmarks by value and delta sizes

This makes it easier to if / when the value or delta increases the
compution time because they each get their own category in the output.

* Reduce the benchmark categories

Else it will run forever...

* A few helpful comments

* Fix spelling

* Have tests cover all cases

This adds tests for missing coverage cases and cleans the tests up a bit
to make it easier to trace failures.

* Setup buffer before benchmark iterations

This removes the buffer setup noise from the actual benchmark run.

* Unpack directly in the decoder

The enum seems to have caused a 15% slow down.

* Simplify header type with more variants

* Make encode easier to read

* Better trait name

* Read delta as u64 directly

* Speed up float read when delta is 0

I don't know why this makes the float reads faster when the delta is 0,
but the C code has the same ordering anyway.

* Add a prop test

* Implement `Seek` for the `BufferWriter`

This is needed to be able to use the `BufferWriter` in the bencher.

* Remove `Copy` from `Decoder::seek`

* Read delta and number in one go

Reducing the number of read from 3 to 2 has about a 10% speedup.

* Reduce bench times

* Give everything consistent names

Even if it is different from the original C names used.

* Don't use the C wrappers for the buffer benches

The final Rust implementation won't use them eiter so it does not make
sense for the benches to be affected by them.

* Leave inlining to the compiler

This does not change the bench results at all.

* Run the bench for shorter

The results seem stable at this run times.

* Remove seek implementation for `BufferWriter`

This is no longer needed and makes the PR simpler.

* Don't set id twice

* Remove duplication in tests

* Address linting issues

* Better bench group names

* Better group names

Co-authored-by: Guillaume Desmottes <gdesmott@gnome.org>

* Use `iter_batched` instead

This properly localizes the overhead of making the buffer.

* Move to own module

This will make it easier to add benchers for other encoders & decoders.

* Add missing license header

* Use `iter_batched_ref`

Without this these iters have a `Drop` call on each benchmark iteration
which slows down the iteration. On some of these that drop call makes it
up to 200% slower.

* Only match on valid types

This will prevent new types slipping through if they are ever added to
the enum.

* Match in the order of most usage

The floats were first for some benchmarking reason, but that is no
longer reproducible.

* Revert to a `&mut Reader` for the decoder

* Make sure all the bytes are written

Unlike the previous call, this continous to call `write_vectored` until
all the buffers have been written out.

* Copy less by using std `From` impl

* Spell fix

* Linter fixes

* Rename `numeric` bench to `encoding_decoding`

* Optimize writing out an array of `IoSlices`

* Expand comment

* Fix spelling

* Don't decode an empty buffer

* Remove no inlining

This was commited by mistake.

* Improve comment

Co-authored-by: Luca Palmieri <20745048+LukeMathWalker@users.noreply.github.com>

* Add `as_numeric` method

This makes it easier to get the numeric data of a record if it is indeed numeric.

* Convert tests into integration tests

* Add property tests for integers

* More tests

* Allow numeric compression

* Fix lint issue

* Spell fix

* Add safety comment

* Inject config for f64 compression

* Improve comment

Co-authored-by: Luca Palmieri <20745048+LukeMathWalker@users.noreply.github.com>

* Make threshold a constant

* Add default impl for linter

* Address lint issue

* Update bencher with constructors

---------

Co-authored-by: Guillaume Desmottes <gdesmott@gnome.org>
Co-authored-by: Luca Palmieri <20745048+LukeMathWalker@users.noreply.github.com>