[MOD-10286] Implement frequency only encoder/decoder in Rust (#6378)
* implement Debug on Buffer and TestBuffer

* factor out buffer.c to its own internal lib

Needed to benchmark C code.

* inverted_index_bencher: test C encoders and decoders

Link on libbuffer as we need Buffer_WriteAt.

* implement freqs only encoder/decoder in Rust

* bench freqs only encoder/decoder

* buffer: gate Debug with debug_assertions

* handle qint_decode failing

* move tests to its own dir

* improve tests

* enforce u32 delta

* remove clone when running Rust test

* stop returning Ok(None) when trying to decode empty buffers

We can't know beforehand if a reader is empty or not. So the only way to
implement this would be to start each decoder by a one byte read which
is suboptimal in most cases.

* buffer.h

* fix test

* update for latest API changes

* Update src/inverted_index/inverted_index.c

Co-authored-by: Pieter <pieter@chesedo.me>

* Update src/redisearch_rs/inverted_index/src/freqs_only/mod.rs

Co-authored-by: Pieter <pieter@chesedo.me>

* move module to a single file

* fix clippy warning

* remove TryFrom<Delta> for u32

Not actually useful as u32 -> usize conversion will always succeed.

The encoder implementation is now properly checking for overflow when
casting Delta to u32.

---------

Co-authored-by: Pieter <pieter@chesedo.me>