Rust FFI cleanup (#6503)
* Define `t_fieldMask` correctly

It should not be hard-coded as a `u128` but should be CPU pointer width
dependant.

* C header file for `low_memory_thin_vec::Header`

We (will soon) have multiple `cbindgen` generated header files depending
on `low_memory_thin_vec`. The problem is that each of these header files
will re-define `low_memory_thin_vec::Header`. So this makes one header
file with the definition which the others can include.