Hide internals of `RSValue`
Ensure code outside of the `value` C module
does not access internals of the `RSValue`
type, but instead uses several newly created
constructors, setters, and getters.

The `RS_DUOVAL_OTHER2VAL` is to be eliminated
in a later commit that renames the `Duo`
variant of `RSValueType` to `Trio`.

`RSVALUE_MAP_KEYPOS` and `RSVALUE_MAP_VALUEPOS`
will be eliminated in a redesign of the map creation
functionality of `RSValue`