Move `RSValue_SendReply` to `reply` module (#6803)
* Move `RSValue_SendReply` to `reply` module

Moves `RSValue_SendReply` to `reply` module, renaming
it to `RedisModule_Reply_RSValue`. Furthermore,
exposes `RSValue_NumToString` from `value.h`,
as it's used in `RedisModule_Reply_RSValue`,
adding a debug assertion that the pased `RSValue` is
indeed a number, so as to make things safer.

* Remove RSValue_SendReply_Collection

Removes `RSValue_SendReply_Collection`, which
was disabled, removing the last reference to
the `reply` module from `value`

* Clarify requirement of RSValue_NumToString parameter being a number

* Use `RS_ASSERT` instead of `assert` in `RSValue_NumToString`

Co-authored-by: GuyAv46 <guy.avimor@gmail.com>

---------

Co-authored-by: GuyAv46 <guy.avimor@gmail.com>