Inverted Index FFI preparation and tweaks (#6851)
* Use getter for inverted index access

* Only set numeric filter if not null

* Numeric encoder can have duplicate entries

* Raw doc id encoder should use first block id for base

* Make helper inline static

This is to get around multiple definitions compile errors.

* Set wide schema flag on tested index

These tests add records with a wide field mask but were not using the
encoders which supports the wide field mask. So this fixes that.

* Update include path

* Have `NewIndexReader` take an owned decoder ctx

* Remove calls to size of ii and index blocks

These sizeof calls won't be possible once these are in Rust.

* Remove access to block size constants

These won't exist in the Rust move. I also don't know why the tests did
not have these hardcoded in the first place.

* Add comment to panic

* Fix spelling

* Fix tiny negatives numeric encoding

These used to be incorrectly encoded as positive tiny numbers.

* Write numeric records to a numeric index

The Rust code will panic without this.

* FFI for IndexBlock

This is only needed by some C tests.

* Fix comment

Co-authored-by: GuyAv46 <47632673+GuyAv46@users.noreply.github.com>

---------

Co-authored-by: GuyAv46 <47632673+GuyAv46@users.noreply.github.com>