More naming consistency, add doc comments to constructors in `value`
Add doc comments to the constructors defined in `value.h`,
clarifying whether they are placed on the stack or on the
heap and whether they should be freed or not.

Apply consistent naming to `RSValue_ParseNumber`,
`RSValue_NewArray`, `RSValue_VStringArray`,
`RSValue_StringArray`, and `RSValue_StringArrayT`
to clarify whether they allocate or not.