[MOD-11587] (1/2) Refactor API of `value` C module (#6956)
* Hide internals of `RSValue`

Ensure code outside of the `value` C module
does not access internals of the `RSValue`
type, but instead uses several newly created
constructors, setters, and getters.

The `RS_DUOVAL_OTHER2VAL` is to be eliminated
in a later commit that renames the `Duo`
variant of `RSValueType` to `Trio`.

`RSVALUE_MAP_KEYPOS` and `RSVALUE_MAP_VALUEPOS`
will be eliminated in a redesign of the map creation
functionality of `RSValue`

* Rename `RSValue` `Duo` type to `Trio`

In the past, `RSValue` has been adapted such
that its `Duo` field actually holds three
values. In preparation for the migration
to Rust of the `value` module, this type
is renamed to `Trio`, eliminating the
`RS_DUOVAL_OTHER2VAL` macro in the process.

* Refactor `RSValue` map API

Add types `RSValueMap` and `RSValueMapEntry`,
representing an `RSValue` map variant and
its entries. Add functions `RSValueMap_Create_Uninit`,
`RSValueMap_SetEntry`, and `RSValue_NewMap`
to allow for creating maps and wrapping them
in `RSValue`s.

* Move `APIVERSION_RETURN_MULTI_CMP_FIRST` from `value.h` to `search_ctx.h`

* Eliminate macros `RSVALUE_REPLACE` and `RSVALUE_CLEARVAR`

* Rename `RSValue_Decref` to `RSValue_DecrRef`

For consistency with `RSValue_IncrRef`,
renames `RSValue_Decref` to `RSValue_DecrRef`.

Also removes and updates outdated comments.