[MOD-11587] (2/2) Further clean up and document the `value` C module  (#6957)
* Naming consistency in `value`

- Rename variants of `RSValueType` and `RSStringType`
  so they are prefixed with their type name.
- Rename constructors to clarify whether they allocate
  the newly created `RSValue` on the heap or not by
  suffixing heap-allocating constructors with `Alloc`,
  and constructors creating a static RSValue-backed
  instance with `Static`.
  Furthermore, ensure prefixing is consistent with
  other functions.

* Reorder functions an definitions in `value` C module

In `value.h` and `value.c`, the function definitions
are ordered and grouped logically: constructors on top,
getters/setters below that, and other functions and
utilities at the bottom.

* More naming consistency, add doc comments to constructors in `value`

Add doc comments to the constructors defined in `value.h`,
clarifying whether they are placed on the stack or on the
heap and whether they should be freed or not.

Apply consistent naming to `RSValue_ParseNumber`,
`RSValue_NewArray`, `RSValue_VStringArray`,
`RSValue_StringArray`, and `RSValue_StringArrayT`
to clarify whether they allocate or not.

* Rename constructors in `value`

To enhance predictibility, the naming scheme of constructors defined
in the `value` C module are renamed:
- heap-allocating constructor names contain "New"`, the "Alloc" suffix is removed
- stack-allocating constructors no longer contain "New"

* Address review comments

- Add doc comments to `RSValue_NewConstString`, and
  `RSValue_NewConstCString`.
- Rename `RSValue_NewRedisString` to `RSValue_NewBorrowedRedisString`
  to clarify the difference with `RSValue_NewOwnedRedisString`
- Have `convert_to_number` in `value.c` return a `bool` instead of `int`
- Rename `RSValue_MallocString` to `RSValue_String` to clarify
  its similarities and differences with `RSValue_NewString`