Link hiredis only if it has been built in sds_rs
There's no reliable way of determining whether we actually
need to build and link to `hiredis` or not
(see <https://github.com/rust-lang/cargo/issues/4001>).
For instance, when running `cargo check`, there's no need to build `hiredis`,
but when building unit tests, `hiredis` needs to be there.
Therefore, we employ a feature flag in the `sds_rs` crate that
controls whether its C dependencies are being linked to and
enable that feature when `value` is built for testing.