Allow RSIndexResult in RQEIterator be mutated when wrapping iterators (#7040)
* Allow RSIndexResult in RQEIterator be mutated when wrapping iterators

Closes MOD-11768

Some RQEIterator implementations wrap one or more other RQEIterator’s.
As part of their read implementation they might however for
some code paths mutate the result by for example
applying a weight to it or metrics.

This was prior to this commit not possible due to lifetime constrains
and the fact that a shared reference is returned.

This commit introduces a Cow-inspired structure to make it easy to facilitate this,
ensuring it is still cheap for code paths where shared references are possible,
but where a wrapper can update to an owned object (derived from child)
where required. E.g. for the optional iterator to apply a weight.

* fix linking errors for simple rqe_iterators test

* switch from Cow-based design for Iterator's result to &mut

* address feedback

* try to inline tmp (re)setters

* remove reset logic, simplifying PR more