Changes in iterator revalidate API (#7135)
* revalidate() returns a Result

Will need to be able to propagate the skip_to() error in the II iterator
implementation.

* return the current document with RQEValidateStatus::Moved

The C API has accessor to retrieve the current document of the iterator
after revalidation. The Rust version does not, so we need to return it
directly from revalidat().

Need to remove Eq as RSIndexResult cannot implement Eq because of its
f64 fields.