[MOD-11174] Port NewInvIndIterator_NumericFull to Rust (#6796)
* iterator: use thiserror

Will allow us to use '?' when reading from the II which is returning a
std::io::Error

* add NumericFull iterator

Rust re-implementation of NewInvIndIterator_NumericFull().
The tests logic have been copied from the existing C++ tests.

Validation is not implemented yet.

* generate bindings for the II C iterator

I had to edit a comment block in tag_index.h to workaround a nasty bug
in bindgen, see https://github.com/rust-lang/rust-bindgen/issues/1313

Also add simple wrappers so it can easily be used in the bencher.

* bench II numeric full iterator

* make iterator test more generic

Will be used to share code with the term iterator tests.

* add NumericDecoder and NumericReader marker traits

Will be used to restrict the numeric II iterator to readers producing
numeric results.

Also allow us to restrict the type of reader wrapper by
FilterNumericReader and FilterGeoReader.

* add FullIterator abstraction

Allow the numeric iterator to be used with a numeric filter wrapper.
Will also be used to implement the term iterator.