Eliminate `RSStringType_Malloc` in `RSValue`
`RSStringType_Malloc` and `RSStringType_RMAlloc` essentially being the same
in terms of freeing strategy, we remove `RSStringType_Malloc` and default
to `RSStringType_RMAlloc` in `RSValue` string constructors instead.

This also does away with the need for taking an explicit string type parameter
in `RSValue_NewStringArrayT` as well as allows us to replace `RSValue_NewStringWithType`
with `RSValue_NewConstString` that does not take a string type parameter.