Control inverted index blocks growth (#7195)
* Control inverted index blocks growth

* Don't compact old blocks

We can't correctly report back the memory saved on this compaction since
these functions return an unsigned integer. And not report this memory
safe correctly causes the index size on stats to no longer be the actual
index size (it has a size bigger than the actual size). So this is
removed for now.

* Lint issue