Switch to rust ii (#6958)
* Switch to Rust inverted index

* Remove encoder and decoder getter usage

It is no longer possible to get the encoders or decoders from the flags.
Instead they are coupled to the inverted index or its reader.

* Update `testNumericInverted`

* Update testIndexFlags

* Don't check buffer growth in test

There is no reason for the test to check the internals of these structures.

* Fix size test on flags

* Fix size test for numeric index

* Fix size test for tag index

* Fix sizes on LLapi

* Update fork tests to const

* C GC updates

* Update fork tests

* Don't free deltas twice

* Fix sizes in Python tests

* Recreate index to have compression turned on

* Remove impossible test

This test requires a mutable access to the inverted index in a reader
and direct access to the repair block call. Both of these are not
possible so the test is removed.

* Remove C benchmarks

* Move header out of future

* Remove from iterator benches

* Fix symbols for GC benchmark

* Update iterator benchmark FFI

* Get num_entries on index directly

* Trim bench dependencies

* Fix filters in CPP tests

The text inverted indices expect to have a field mask filter, not the
`None` case.

* Update debug assertions

It is now possible to have an inverted index with no blocks in it.