Generalize opaque sized type pattern (#7234)
* Generalize opaque sized type pattern

In order for types defined in Rust to be placed on the
stack in C, they need to be sized and FFI-safe. Annotating
these types with `#[repr(C)]` makes the type's layout well-defined,
but comes with two downsides:

1. (Private) fields are exposed to C code, allowing C code
   to rely on the internals, and even break invariants.
2. All fields or variants of the type need to be FFI-safe. Thefore,
   types like `String` and `Arc` cannot be used.

The `query_error_ffi` crate instead introduces an opaque sized type,
`OpaqueQueryError`, which is validated to have the same size and alignment
of `QueryError`, but hides its internals and is FFI-safe even if
`QueryError` is not.

This commit extracts that implementation into a trait called
`IntoOpaque`, and introduces a macro `opaque` that makes
implementing `IntoOpaque` less error-prone, by doing some
compile-time assertions. It also implements `IntoOpaque`
for `RsValue`, and uses it in `value_ffi`. `RsValue`
is no longer `#[repr(C)]`.

Due to limitations of cbindgen, `opaque` needs the user to
define the opaque type themselves, but the compile-time
assertions and safety requirements make this rather trivial.

* Implement `IntoOpaque` for `QueryError`, remove `query_error_ffi::opaque` module

* Add license headers and clarify docs