[MOD-11717] (1/3) Combine behavior of `RsValue` and `SharedRsValue` trait `Value` (#7322)
* Add `expect_unchecked` macro.

Introduce `expect_unchecked` macro, which is to be invoked with Options and Result, that expands to a call to `expect` in debug mode, and a call to `unwrap_unchecked` in release mode. That way, invariants can be checked during tests, while being skipped entirely in optimized builds.

* Provide access to `RSDummyContext` in Rust

In `src/module-init`, the value `RSDummyContext` is set upon module initialization, which is required for manipulating Redis strings, among others.

* Add `Value` trait to implement shared behavior of `RsValue` and `SharedRsValue`

Implement it on those types.

`RsValue` and `SharedRsValue` differ in allocation strategy, but not in behavior. The `Value` trait provides the implementation of the shared behavior.