[MOD-12360] move FieldMaskOrIndex to Rust (#7332)
* Turn FieldMaskOrIndex to a union

This match the type that will be generated by cbindgen when moving
it to Rust as an enum.

The size of FieldMaskOrIndex is unchanged as both versions use 32 bytes:

The old version embedding an union into a struct:
- bool isFieldMask: 1 byte
- Padding: 15 bytes
- union value containing either t_fieldMask (16 bytes) or t_fieldIndex (2 bytes)
Total: 32 bytes

The new pure union version:
- FieldMaskOrIndex_Tag tag: 1 byte (uint8_t)
- Or a struct with tag (1 byte) + t_fieldMask index (16 bytes)
- Or a struct with tag (1 byte) + t_fieldIndex mask (2 bytes)
Total: 32 bytes (with padding)

* move FieldMaskOrIndex to Rust

Will be needed to implement II query iterators.