Introduce DynRsValue and DynRsValueRef, expose them to C
This commit introduces `DynRsValue` and `DynRsValueRef`.
`DynRsValue` is an enum holding either an `RsValue` or a `SharedRsValue`.
It allows for creating an FFI module that is very similar to the C
`value` module, in which a single `RSValue` type is exposed.
Using `DynRsValue`, there is no longer need for exposing methods
for both `RsValue` and `SharedRsValue`.

`DynRsValueRef` serves as a cheap-to-dereference reference to either
an `RsValue` or a `SharedRsValue`.

`RsValue` and `SharedRsValue` still are separate types on the Rust side,
allowing Rust to clearly differentiate between refcounted instances and
ones that are not refcounted.