Don't access `NULL` GC delta (#7547)
* Explicitly set delta to NULL after free

* Simplify checks around free GC delta

Since we are setting the pointer to NULL we no longer need the boolean
check. On the `InvertedIndex_GcDelta_Free` call, Rust will only free the
delta if it is not NULL.

* Don't access `NULL` GC delta

The GC delta was freed from memory when it was applied a few lines
before the `resetCardinality` call. This changes gets the last block ID
in the delta before this apply call which frees it, and then passes this
last block to `resetCardinality` instead.

* Reset old logic for finding (new) last block