Rust CI improvements (#7520)
* Move nightly version to a standalone .rust-nightly file.

All nightly-related references to build.sh have been updated
to use .rust-nightly as the source of truth.

* Copy `.rust-nightly` rather than `build.sh` into the CI container.

It gets rid of unnecessary rebuilds, since the nightly version was
the only thing we cared about from `build.sh` when building the container.

* Use a specific stable version of Rust.

Using a specific version, the CI container is correctly rebuilt whenever that version changes.
It also moves the Rust toolchain file to the root of the repository, to ease
copying it into the container during build.

* Cache Rust packages in CI.

* Only use the cache action if Node is supported.

* Trigger build with cache hit