fix: replace interiour NULL bytes with spaces.
This fixes a potential issue where the logger would stop in the middle of a record if that record contains null bytes (which are valid UTF8) because it treats it as the end of the C string.
We check (and potentially replace with a space character) all characters in the record. This loop should hopefully be trivially vectorized.