feat: `tracing-redismodule` (#7548)
* feat: `tracing-redismodule`

This change implements a subscriber for the Rust `tracing` ecosystem that emits traces and logs to the redismodule logging system.

In its current state kept intentionally simple, it allows configuring the logging output via the standard `RUST_LOG` environment variable
but not much more.

* support configuring output styling

* correctly set redismodule log level

* add usage docs

* fix: replace interiour NULL bytes with spaces.

This fixes a potential issue where the logger would stop in the middle of a record if that record contains null bytes (which are valid UTF8) because it treats it as the end of the C string.
We check (and potentially replace with a space character) all characters in the record. This loop should hopefully be trivially vectorized.

* tests: add python tests for tracing_redismodule

Adds a few python integration tests to ensure basic operation of the tracing_redismodule subscriber (we can set level and crate filtering through the env var)

* fix: remove duplicate entries